Folder restructuring as shared,web adn native
