+++ index.android.js
@@ -8,15 +8,15 @@ import React, { Component } from 'react';
 import { AppRegistry } from 'react-native';
 import { StackNavigator } from 'react-navigation';
 
-import Teams from './src/views/TeamsListView';
-import Login from './src/views/LoginView';
-import MainCategory from './src/views/MainCategoryView';
-import ManageCase from './src/views/ManageCaseView';
-import Preferences from './src/views/PreferencesView';
-import SetInfo from './src/views/OfficeCareView';
-import AddPreference from './src/views/AddPreferenceView';
-import OfficeSets from './src/views/OfficeSetsView';
-import CasesDayView from './src/views/CasesDayView';
+import Teams from './src/views/native/TeamsListView';
+import Login from './src/views/shared/LoginView';
+import MainCategory from './src/views/native/MainCategoryView';
+import ManageCase from './src/views/native/ManageCaseView';
+import Preferences from './src/views/native/PreferencesView';
+import SetInfo from './src/views/native/SetInfoView';
+import AddPreference from './src/views/native/AddPreferenceView';
+import OfficeSets from './src/views/native/OfficeSetsView';
+import CasesDayView from './src/views/native/CasesDayView';
 
 const OptimizeApp = StackNavigator({
     Login: { screen: Login },
