+++ package.json
@@ -11,6 +11,7 @@
     "react": "15.4.2",
     "react-native": "0.42.3",
     "react-native-alphabetlistview": "^0.2.0",
+    "react-native-atoz-listview": "0.0.7",
     "react-native-calendar": "^0.11.0",
     "react-native-modal-dropdown": "^0.4.2",
     "react-native-selectme": "^1.2.3",
