+++ src/components/native/case-view/style.js
