+++ src/components/native/footer/index.js
