+++ src/components/native/footer/style.js
