+++ src/components/native/opacity-btn/style.js
