+++ src/components/native/ticker/style.js
