+++ src/components/shared/text/index.js
