+++ src/components/shared/view/Render.js
