+++ src/components/shared/view/Render.native.js
