+++ src/components/shared/view/index.js
