+++ src/json/setInfo.json
@@ -1,28 +1,20 @@
 {
-    A: ['some', 'entries', 'are here', 'some', 'entries', 'are here'],
-    B: ['some', 'entries', 'are here'],
-    C: ['some', 'entries', 'are here'],
-    D: ['some', 'entries', 'are here'],
-    E: ['some', 'entries', 'are here'],
-    F: ['some', 'entries', 'are here'],
-    G: ['some', 'entries', 'are here'],
-    H: ['some', 'entries', 'are here'],
-    I: ['some', 'entries', 'are here'],
-    J: ['some', 'entries', 'are here'],
-    K: ['some', 'entries', 'are here'],
-    L: ['some', 'entries', 'are here'],
-    M: ['some', 'entries', 'are here'],
-    N: ['some', 'entries', 'are here'],
-    O: ['some', 'entries', 'are here'],
-    P: ['some', 'entries', 'are here'],
-    Q: ['some', 'entries', 'are here'],
-    R: ['some', 'entries', 'are here'],
-    S: ['some', 'entries', 'are here'],
-    T: ['some', 'entries', 'are here'],
-    U: ['some', 'entries', 'are here'],
-    V: ['some', 'entries', 'are here'],
-    W: ['some', 'entries', 'are here'],
-    X: ['some', 'entries', 'are here'],
-    Y: ['some', 'entries', 'are here'],
-    Z: ['some', 'entries', 'are here'],
+    "A": [{
+            "name": "Anh Tuan Nguyen",
+            "age": 28
+        },
+        {
+            "name": "An Nguyen",
+            "age": 20
+        }
+    ],
+    "Z": [{
+            "name": "Zue Dang",
+            "age": 22
+        },
+        {
+            "name": "Zoom Jane",
+            "age": 30
+        }
+    ]
 }
