+++ src/views/native/AddPreferenceView/style.js
