+++ src/views/native/CasesDayView/style.js
