+++ src/views/native/MainCategoryView/style.js
