+++ src/views/native/ManageCaseView/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import NavBar from '../../components/navbar';
-import TouchableOpacityBtn from '../../components/opacity-btn';
-import Calender from '../../components/calender';
+import NavBar from '../../../components/native/navbar';
+import TouchableOpacityBtn from '../../../components/native/opacity-btn';
+import Calender from '../../../components/native/calender';
 import {
   StyleSheet,
   Text,
