+++ src/views/native/ManageCaseView/style.js
