+++ src/views/native/SetInfoView/style.js
