+++ src/views/native/TeamsListView/style.js
